Move Header into separate component
